make the map default page
